Incorporate QModel v3 and move files to subfolders
