--- QATCH/QModel/__init__.py
@@ -1,2 +1,0 @@
-__version__ = "2.0.12"
-__release__ = "2025-02-17"
