+++ QATCH/QModel/src/models/static_v2/__init__.py
@@ -1,0 +1,2 @@
+__version__ = "2.0.12"
+__release__ = "2025-02-17"
