+++ QATCH/QModel/src/models/static_v3/__init__.py
@@ -1,0 +1,2 @@
+__version__ = "3.0.0"
+__release__ = "2025-04-18"
